+++ src/components/select/multi-select/multi-select.test.js
@@ -966,7 +966,7 @@ context("Tests for Multi Select component", () => {
 
         dropdownButton().click();
         selectListPosition()
-          .should("have.attr", "data-popper-placement", flipPosition)
+          .should("have.attr", "data-floating-placement", flipPosition)
           .and("be.visible");
       }
     );
@@ -974,24 +974,18 @@ context("Tests for Multi Select component", () => {
     it.each([
       ["bottom", "0px", "0px", "0px", "0px"],
       ["top", "600px", "0px", "0px", "0px"],
-      ["right", "200px", "0px", "0px", "900px"],
-      ["left", "600px", "0px", "900px", "0px"],
+      ["bottom", "200px", "0px", "0px", "900px"],
+      ["top", "600px", "0px", "900px", "0px"],
     ])(
-      "should render list in %s position with the most space when listPosition is set to auto",
+      "should render list in %s position with the most space when listPosition is not set",
       (position, top, bottom, left, right) => {
         CypressMountWithProviders(
-          <MultiSelectComponent
-            listPlacement="auto"
-            mt={top}
-            mb={bottom}
-            ml={left}
-            mr={right}
-          />
+          <MultiSelectComponent mt={top} mb={bottom} ml={left} mr={right} />
         );
 
         dropdownButton().click();
         selectListPosition()
-          .should("have.attr", "data-popper-placement", position)
+          .should("have.attr", "data-floating-placement", position)
           .and("be.visible");
       }
     );