+++ src/components/select/simple-select/simple-select.component.js
@@ -51,7 +51,7 @@ const SimpleSelect = React.forwardRef(
       "data-component": dataComponent,
       "data-element": dataElement,
       "data-role": dataRole,
-      listPlacement = "bottom-start",
+      listPlacement = "bottom",
       flipEnabled = true,
       ...props
     },
@@ -472,23 +472,7 @@ SimpleSelect.propTypes = {
   /** Maximum list height - defaults to 180 */
   listMaxHeight: PropTypes.number,
   /** Placement of the select list in relation to the input element */
-  listPlacement: PropTypes.oneOf([
-    "auto",
-    "auto-start",
-    "auto-end",
-    "top",
-    "top-start",
-    "top-end",
-    "bottom",
-    "bottom-start",
-    "bottom-end",
-    "right",
-    "right-start",
-    "right-end",
-    "left",
-    "left-start",
-    "left-end",
-  ]),
+  listPlacement: PropTypes.oneOf(["top", "bottom", "right", "left"]),
   /** Use the opposite list placement if the set placement does not fit */
   flipEnabled: PropTypes.bool,
 };